Refactor away org & space providers / decorators
Since identity is now determined on demand in the namespace permissions
object, there is no need for an org or space provider to get the
identity and inject it into the decorators. This makes the provider
redundant.

Further, the decorators where decoratoring an 'auth disabled' list
orgs/spaces method, where a privileged client returned orgs and spaces
regardless of the user's permissions. This will be deleted soon, so the
decoration was the wrong way round.

Instead of inverting the decorator, which would have involved
replicating some logic, we decided to pass an authEnabled flag into the
org repo, and when that is not set, we just return early from the list
orgs/spaces methods without filtering the results based on the user's
role bindings.

This is the final commit after tidying up the changes made to list apps
efficiently.

Issue: https://github.com/cloudfoundry/cf-k8s-controllers/issues/158
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>