Cache identities in new  CachingIdentityProvider
Use an expiring cache with TTL of 2 minutes.

The identity provider is used to determine the ID given an auth token or
client certificate. In both cases, the code goes to the cluster to
obtain the identity, and this is a slow operation. It makes sense to
cache the results for a short time.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>