Add an admin user on kind deployment
Use the new scripts/create-new-user.sh script to create a
user with name admin and bind this user to the cf-k8s-controllers-admin
role. This role is a clone of the cf-k8s-controllers-space-developer
to begin with and should be incrementally granted more permissions as we
add new endpoints to the api. The goal is to eventually have and admin
role that is just as powerful as needed, but not as powerful as the k8s
admin user.

Issue: https://github.com/cloudfoundry/cf-k8s-controllers/issues/466