Authorize GET /v3/apps/{guid}
At the moment we still retrieve the app by guid in an inefficient way,
going through the entire list of apps retrieved by a privileged client
looking for the matching guid. Then we use the user client to perform a
`Get` once we know the namespace and name.

This need to be fixed in a generic way using some kind of cache.

Issue: https://github.com/cloudfoundry/cf-k8s-controllers/issues/234
Co-authored-by: Kieron Browne <kbrowne@vmware.com>