Ensure correct authz failure for get current droplet
This adds an e2e test to create a droplet and assign it on an app. That
entails creating an app, creating a package associcated with the app.
Uploading source to the package and creating a build associated with the
package. Waiting for the build to complete then assigning the current
droplet on the app.

The authz is actually taken care of when getting the app in this case,
so no repo code has changed. We did fix the error handling in the app
handler though.

Issue: https://github.com/cloudfoundry/cf-k8s-controllers/issues/440
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>