E2E tests wait for admin role binding in orgs/spaces
The org repo waits for the subnamespaces anchor status to become OK,
unfortunately this only indicates that the underlying namespace has been
created. Objects propagation is carried out asynchronously, therefore
there is a tiny amount of time that rolebindings from the parent
namespace are not available into the child namespace. During this short
interval the org/space might be unusable even for admins as they are not
yet bound to their roles in that particular namespace.

HNS currently does not provide a mechanism to signal that propagation is
complete so it is very hard for us to implement a proper fix. Therefore
until we are sure that this is an actual problem, e2e tests would just
wait for the admin role binding to appear in the namespace which should
be good enough to avoid flakes.

Co-authored-by: Mario Nitchev <marionitchev@gmail.com>