Fix line numbers in ginkgo error output in E2E tests
Ginkgo's default fail handler (the one the E2EFailHandler delegates to)
uses the `callerSkip` argument to skip stack frames when printing the
error line. As we wrap the standard fail handler with a custom one, we
should increase the `callerSkip` argument with one (wrapping the default
handler adds one stack frame) so that the error message points to the
actual failure.

This fix also works well with `ExpectWithOffset`. Being on the subject,
the offsets in `createRole` have been adjusted to `2` which is the
correct value.