Integrate the run SupplyChain with ConventionService
The run SupplyChain now creates PodIntents, which the ConventionService
modifies (by default it's a no-op), and the result is used to deploy a
StatefulSet.

This change removes Eirini integration, since Eirini doesn't accept pod
specs, and that's what the ConventionService acts upon.

We attempted to change the run SupplyChain into a ClusterDelivery, but
Deliveries cannot use ClusterConfigTemplates, which are how we create
PodIntents and feed their output into the StatefulSet. If Deliveries
support ClusterConfigTemplates at a later date, then we could attempt
this again.

Co-authored-by: Matt Royal <mroyal@vmware.com>
Co-authored-by: Julian Hjortshoj <hjortshojj@vmware.com>