Remove domain creation and deletion from e2e suite
Move it to run-tests.sh. The default domain will be created there if
APP_DOMAIN_GUID is not set.

This has required a change in the route e2e test. We now test for an
invalid host name rather than an invalid FQDN. The latter required an
invalid domain name, which really shouldn't be possible via the API.

The domains listing tests in org and domains files have been amended to
check for returning the default domain only. This can be modified to
return more than one if required once the create and delete domain API
endpoints are available.

Issue: https://github.com/cloudfoundry/cf-k8s-controllers/issues/830