Fix e2e list flakes for app and service binding
Some app and service binding tests checking list results need to account
for apps and service bindings created by other tests. So instead of
checking for an absolute length of the list results, check that expected
elements are contained, and unexpected elements are not contained. That
is sufficient in an e2e test.