Setup a commont test org and use it in most tests
Now that we have a shared org that all tests create spaces in
it is very important to delete spaces after the test is done.
If we do not do that HNC becomes orverwhelmed quickly and tests
start flaking. This commit adds space cleanup to all existing files,
but we will have to be mindful of doing that to pervents tests from
dragging in the future