Bump HNC to v1.0.0
When taking the new `default.yaml` for v1.0.0 from the hnc release, I
realised it could be easy to lose the manual modifications we had made
to the previous `hnc-manager-v0.9.0.yaml` file. So I've put the changes
in a patch file to apply with kustomize to make this more secure.

The bash lines to wait for HNC to come up fully have also been removed
as v1.0.0 has a proper /readyz implementation.

Fixes: https://github.com/cloudfoundry/cf-k8s-controllers/issues/908