Tweak waiting for HNC propagation
We have reverted the HNC qps setting to 50 as 200 seems to make GKE
cluster API unresponsive, causing nodes to be recreated.

We have also made the timeout in the org repo longer. It uses the same
timeout as used for waiting for the subnamespaceanchor to be ready.

Since the average time for propagation to complete is around 8 seconds
during heavy load on GKE during e2e tests, we have decided not to use
exponential backoff. This was missing a check at around 6 seconds then
having to wait until 12 seconds to succeed. Simply checking every 500ms
seems to provide better results.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>