Remove the `Serial` decorator from ListServiceBindings test
The repository is using a unprivileged client so the repo cannot see
service bindings created in unauthorised namesaces. Therefore using the
`Serial` decorator when listing packages does not make sense.

While being here, the code under test has been moved to `JustBeforeEach`
to remove duplicate invocations in the `It`s