Update exploration issue template
We talked as a team about how we've been using the "days to checkin" discussion during IPM just to discuss complexity and not to actually determine a timebox. Updating this template to make it more clear that this is optional and can be included if we decide an explicit timebox is helpful.