Remove Eventually statements from api related tests
Some Eventually's that make sense have been left in place. This is
usually the case for resources that are not directly created by the
client (e.g. process, droplet, etc). When the API has async contract by
returning a 202 with a Location header, the same applies. In all other
cases we assume that resources should be directly available to the
client.

The goal is to expose any potential flakes due to the async nature of
k8s, evaluate how critical they are and decide how to best handle them,
rather than the current approach of hiding this information.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>