Trim permissions of the api shim service account
This commit removes all unneeded kubebuilder annotations and updates the
existing ones to reflect what is currently needed by the code.

We also needed to fix the role repository to remove the privileged
client. Previously it required that for a namespace lookup, but that can
be handled properly by the org repo, so we injected a dependency for
that.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>