Fix codeclimate links in readme
This was failing after the cf -> korifi rename.

I've recreated the codeclimate repo analysis linked to my kieron-dev
github account. It's not a problem that it's tied to this account, as
anyone else can add the repo to theirs in the future if required. That
would supersede my one.

The code coverage test reporter ID is now stored in a github secret.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>