Introduce correlation IDs to requests and logs
If X-Correlation-ID is passed in the request headers, this will be used
as an ID for the life of the request. If not passed, one will be
generated. The value is returned in the response headers so a generated
value can be parsed by the calling application if required.

All logs will include the id in the `correlation-id` value.

Removing the development: true flag to the logger, as it's harder to
read with all the multi-line stacktraces in the way, and harded to link
to the correlation ID.

Issue: https://github.com/cloudfoundry/korifi/issues/656
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>