Listing droplets and processwes must tolerate 403 errors
The list of namespaces we list a resource in is just a list of namespace
containing a rolebinding for the current user. This does not mean the
user will be authorized for the list action we are attempting in the
namespace. Therefore we should silently ignore forbidden errors during
list.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>