tarball.LayerFromReader is deprecated
This replaces `tarball.LayerFromReader` from `go-containerregistry`
with `tarball.LayerFromOpener`, as the former is deprecated.

`LayerFromReader` opens the contents of the input `io.Reader` into a
temporary file, before calling `LayerFromFile` which in turn called
`LayerFromOpener`, which then opened the file lazily via the provided
opener lambda.

Our code did in turn copy the contents of the uploaded file into a
temporary file, opened it, wrapper the resulting `io.Reader` with a
ZIP-to-TAR converter and passed that to `LayerFromReader`. It then
deferred the deletion of the temporary file.

This meant we had two temporary files: ours, which we immediately
deleted after the call to `LayerFromFile`, and the library one which
never got explicitly deleted (but should automatically be cleaned up by
the OS).

Now that we call `LayerFromOpener` directly, the library doesn't create
this extra temporary file for us anymore, and we are now in charge of
keeping our temporary file on disk until the opener lambda gets invoked,
which is the first time the layer will be accessed.

I don't believe "leaking" our temporary file should be a problem, as
that's what temporary files are designed for anyway, and we still did it
previously, just without knowing, as the "leaking" was happening inside
the library.