Job handler no longer requires IDs contain a GUID
- We previously required that Job identifiers adhered
to a strict structure of `<operation>-<optional prefix>-<GUID>`.
- This change loosens this restriction so that users using
the CRDs directly are not required to use a GUID and instead are able
to use any valid Kubernetes resource.
- Changes the delimiter between job operation and resource
identifier to be "~". The "~" character is not used by any CF Job
operation types, is a valid non-reserved URL character, and is not a
valid character for Kubernetes resource names. This allows us to
simplify the regular expression and have fewer assumptions around
resource naming.
- Refactors job redirect URL building by centralizing the logic in
the Job Presenter
- Issue: https://github.com/cloudfoundry/korifi/issues/1116

Authored-by: Tim Downey <tdowney@vmware.com>