Allow running condition to be not present
IsStatusConditionFalse only returns `true` when the condition both
exists and is equal to "False". In this test, we need to accommodate the
condition not yet existing also. So the correct way to call this is to
check for the "True" condition, and assert that we don't get it.

Fixes flake at cfapp_controller_integration_test.go:100.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>