Introduce the `Staged` condition for `CFApp`
This adds a `Staged` condition to `CFApp` which will be set to true only
if the app has a valid droplet reference. The condition will be unset
back to false if the droplet referenced by the app disappears.

This will make it very easy for the task creation endpoint to identify
apps that are not staged yet and return proper errors.

Co-authored-by: Danail Branekov <danailster@gmail.com>
Issue: https://github.com/cloudfoundry/korifi/issues/1266