Fixes the fololowing flake
https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-e2es-main/builds/109#L62c509cc:254:286

The VCAP_SERVICES env var is coming from a secret that might not have
been set at the time of query. This fix is waiting for the VCAP_SERVICES
key to appear before inspecting its value. While this fixes the flake is
this behaviour desired? This means that users will have to retry until
they finally see their service instance related env vars