Attempt to de-flake CFSpace Validator int. tests
See the comment we added, but in summary we don't believe the flake will
ever happen in practice and only happens due to the test setup.

To replicate the situation that will happen in production, we ensure
that the CFOrg is present in the client cache used by the validator
itself.

We were never able to get the test to flake, but it continues to pass
locally.

[#1303]