Await the `Staged` app condition on set app droplet
By doing that we ensure that the app controller has accepted the request
to set the app droplet and it is aware that the app is staged. Thus
subsequent creation of objects that depend on the app being staged (such
as tasks) would not fail.

Awaiting for a status condition of a task requires adding the `watch`
permission on apps for the `cf_admin` and `space-developer` roles

Issue: https://github.com/cloudfoundry/korifi/issues/1385
Co-authored-by: Kieron Browne <kbrowne@vmware.com>