Use generics to simplify the awaiter
Note that the 3rd type parameter is required so that we can instantiate
the pointer type privately inside the AwaitCondition method. There is
some go boilerplate to accept the plain type which can be instantiated
and the extra constraint that its pointer must implement the
client.ObjectList interface.

Issue: https://github.com/cloudfoundry/korifi/issues/1385
Co-authored-by: Kieron Browne <kbrowne@vmware.com>