Remove test consistently failing on concourse
The test being removed is able to pass loclally as well as in gh
actions, but fails consistently in our Concourse, because our test are
running as `root` in the ci image.

There seems to be no easy way of fixing this so we remove it to unblock
ci.

Running our tests as non-root is certainly a possibility, but we are not
convinced that it is worth the cost.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>