Set task lifecycle launcher in cftask reconciler
The CFTask resource should mirror the api in accepting a string, rather
than a slice of strings, for its command. So we change CFTask.Command to
the string type.

The responsibility for inserting the CNB launcher into the command then
belongs to the CFTask reconciler, so we move the logic from the task
repo to the CFTask reconciler.

Issue: https://github.com/cloudfoundry/korifi/issues/1430