Fix bug reading stats for recently-scaled-down process
During a call to `cf scale -i`, the CLI first scales down the process
and then fetches its status. At that moment there are likely "extra"
pods, since the statefulset hasn't finished scaling down. This was
leading to a panic as we tried to modify a record at a non-existent
index PodRepo.ListPodStats.