Overlay overrides configmap with desired fields
In the previous commit the overlay introduces a new configmap containg
the desired configuration properties. This has the drawback that this
configmap is created in the default namespace and does not share the
common name prefix withe the rest of the resources.

In this commit the base defines the configmap with reasonable defaults
and the overlay overrides only what's necessary. The configmap holding
korifi's deploy config is in the korifi-api-system namespace and shares
the same name prefix. The substitutions do not need to guess the name
and namespace of the source configmap anymore, which means that the
namespace and prefix are only defined once

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>