Do not ignore droplet process command on start
Processes created during manifest application used to ignore the command
coming from the droplet if any. This lead to processes without a
command, which resulted in the container entrypoint taking over. This
was accidentally working for the `web` process (as CNBs set it as the
main image command) but not correct in general.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>