Allow scaling web processes down to 0
As zero is a valid value for `CFProcess.Spec.DesiredInstances` (set to
zero when the user wants to scale down process instances to 0) we cannot
rely on the field zero value. That is why the field has its type changed
to `*int` so that the CFProcess webhook would only default the instances
value if the pointer value is not set.

Issue: https://github.com/cloudfoundry/korifi/issues/1672