Accept disk-quota with a hyphen in manifests
At both app and process level, `disk-quota` is now accepted, but is marked
as deprecated. `disk_quota` is the correct value to use according to the
docs.

It is an error to use both `disk-quota` and `disk_quota` at either app or
process level.

Issue: https://github.com/cloudfoundry/korifi/issues/1584
Co-authored-by: Kieron Browne <kbrowne@vmware.com>