Make sure CRDs are always installed, updated and deleted
Helm v3 has a policy of not touching CRDs once they are initially
installed. This is problematic for us, as we generally want them to be
updated on the pr-e2e cluster, say, when a CRD definition changes.

This means we cannot use the helm strategy of placing CRDs in the helm
crds directory.

Instead, we include CRDs in the controllers templates. This way they
will always be created / updated / deleted as expected.

A side-effect is that we cannot include the kpack BuilderInfo resource
in the kpack templates. The CRD is generally not available in time for
this resource to be applied. So we have moved its creation and deletion
into helm post-install and pre-delete hooks.

Co-authored-by: Danail Branekov <danailster@gmail.com>