Use the helm release namespace for a delete job
We have seen this delete hook failing when the install hasn't completed
as far as creating the korifi-kpack-image-builder-system namespace. So
it is safer to use the helm release namespace, defaulting to `default`,
which stands a better chance of existing.

Also use a dedicated service account for the deletion to avoid
dependencies, and fix ordering. (Helm priority order is ascending, e.g.
-10, -5, 0, 10)