Issue a warning when the user is not a CF user
Introduce a new `CFUserMiddleware` that checks whether the user has a
rolebinding in the CF root namespace. If not, then set the
`X-Cf-Warnings` header with a value explaining that the subject has
no CF roles assigned. In order to be as transparent as possible (i.e.
not to cause unwanted failures), the middleware would always delegate to
the next handler, regardless what errors it faces.

The warning itself is displayed by the CF CLI which would warn users who
are using the default cluster admin user before explicitly
creating proper role bindings, which has been a great source of
confusion so far.

The middleware uses an expiring cache to cache known CF users. Stale
cache is not an issue here as if the user is no longer a CF one, the
call would be stopped by RBAC anyway.

This reverts commit f72bea499dd4cc4971b2113576fcd45ca3bca8ef.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>