Get rid of the service binding finalizer
The finalizer purpose was to ensure that a service binding is deleted
when an app is being deleted. This has now been replaced with an owner
reference from the app to the binding so that K8S takes care of that for
us. It is the reponsibility of the service binding creator to set the
owner ref, therefore this code was moved to the repository.

The binding finalizer also had the reponsibility to update the app
VCAP_SERVICES secret. Now this is taken care of by the app controller
that now listens to service binding events and rebuilds the secret
accordingly.

While being here, use the awaiter utility to wait for the binding ready
condition in the repository.

Issue: https://github.com/cloudfoundry/korifi/issues/1716
Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>