Remove the finalizer from the CFServiceInstance controller
What the finalizer does is to delete the service bindings related to
that service. This behaviour is now achieved by the service instance
"adopting" the service binding in the service binding controller by
setting an owner reference from the service instance to the binding.
Thus whenever the service is deleted, K8S would delete the service
binding for us.

However, service bindings have to be deleted when the app being bound to
the service is deleted as well, this has not been supported so far. In
order to achieve this, the CFApp finalizer now lists and deletes
associated service bindings. This is however a "best effort", it is
possible that the finalizer misses recently created bindings. We do not
expect this situation to happen in real life, but if it does, then the
orphan binding would remain behind. It would be eventually deleted when
the service instance is deleted.

The orphan service binding should not be much of a problem though as it
references apps by GUIDs, i.e. another app would not be able to reuse
it by accident.

Issue: https://github.com/cloudfoundry/korifi/issues/1876
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>