Commit immediately after setting finalizer in controllers
There is a potential race when deleting a resource with a finalizer
when the finalizer is not immediately persisted. E.g. for a CFSpace:

1. Create the CFSpace
2. Wait for namespace to appear
3. Delete the CFSpace.

However, the reconciler could well be in the code between creating that
namespace and propagating service accounts / secrets, and nothing has
yet been saved.

When the big 'patch' call finally occurs, the CFSpace has already been
deleted, since it didn't actually have a saved finalizer on it.

All resources created by the reconciliation are now orphaned and will
not be cleaned up.

(We also tidied up logs in the touched files)

Issue: https://github.com/cloudfoundry/korifi/issues/1930
Co-authored-by: Danail Branekov <danailster@gmail.com>