Introduce global.containerRegistryCACertSecret
This allows users to specify a CA cert for the container registry. It
also gets rid of the `REGISTRY_CA_FILE` environment variable and of the
custom transport construction we had in place: instead we just mount the
cert unser `/etc/ssl/certs`, which is one of the dirs automatically
scanned by Golang when building its system CAs pool.

Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>