Remove gorilla/schema
This replaces the `gorilla/schema` with a custom implementation. It
consists of adding a `DecodeFromURLValues` to every payload type, and
use it in `payloads.Decode`.

This also gets rid of pointer fields on payload structs, as we didn't
find any reason to have them. In some places we were treating a list
parameter as an empty slice when not passed, and as a slice containing a
single empty string when passed empty: we didn't think this distinction
was useful, so we treat both cases as empty slices now.

Co-authored-by: Dave Walter <walterda@vmware.com>
Issue: https://github.com/cloudfoundry/korifi/issues/2010