Stabilize app revision
The app revision used to be set/incremented un the cf app webhook
according to nontrivial logic that was trying to guess whether the app
is being stopped which was error prone and unpredictable because there
are no guaranteed how many times the hook will execute

This WIP is an attempt to move the logic from the webhook to the app
handler, or to be more exact, to eliminate it, since the app handler
knows with certainty if the app is being stopped or started. The handler
will bump the app  rev when the app stop endpoint is triggered. This
makes the app rev predictable by the user, since it no longer happens in
the reconciliation loop.

Co-authored-by: Danail Branekov <danailster@gmail.com>