Add `--wait` to helm install in docs
There is an obscure dependency between post-install templates and the
`--wait` flag. Post-install templates only wait for the main
installation to be _ready_ if the `--wait` flag is specified in install
or upgrade command.

Our scripts all use `--wait`, but our docs did not reflect that.

Issue: https://github.com/cloudfoundry/korifi/issues/2045