Don't zip dora in when bumping vendir deps
Git doesn't store timestamps, so the vendir'ed code will also have
current timestamps on files, which means the zip timestamps will always
differ from the previous. Hence the korifi-bot will try to PR a dora.zip
change every time it runs.

We leave the current dora.zip for now, and this will be addressed by a
work on #1146 which should generate app zips on the fly.

Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>