Only clear labels/annotations when set to nil
The previous behavior was incorrect. Leaving a value as "blank" in yaml
sets it to nil, not empty string. This implementation allows users to
explicitly give a label or annotation a value of empty string.

Nil values are ignored when creating a new app. For existing apps, a nil
value will remove the matching label/annotation if it exists.

[#1945]