CFPackage uses status conditions
Set initialized condition to true when:
- resource has finalizer set, and
- resource has app owner ref set.

Set ready condition to true when:
- resource has source package image set

Wait for the initialized condition in the CFPackage repo when createing
a CFPackage.

Wait for the ready condition in the CFPackage repo when updating a
CFPackage source image.

This will be used in the story below, where we need to expire old
packages in controllers on the edge where the source image is first set.

Issue: https://github.com/cloudfoundry/korifi/issues/2232
Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>