Cleanup old packages in package reconciler
Here we also stop using the patching reconciler in the package
reconciler as it produced flaky behaviour. Instead we save status or the
main resource as required.

This introduces a new config value `MaxRetainedPackagesPerApp` in the
controllers config, and a corresponding value in the controllers section
of the helm values `maxRetainedPackagesPerApp`.

Issue: https://github.com/cloudfoundry/korifi/issues/2232
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>