CFApp should own CFBuild, not CFPackage
If package owns build, builds get deleted when packages are deleted.
This causes problems when older packages are garbage collected. They
might own non-succeeded builds which we do not expect to be garbage
collected.

Nothing depended directly on controller ownership of build by package.
By adding non-controller ownership of the build by the app, we preserve
auto-deletion when the app is deleted, and also allow seamless migration
for existing builds.

Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>