Set current creation date on kpack build images
We are seeing flakes in google artifact registry where sometimes droplet
pushes fail due to 'config blob not found'.

This _might_ be due to us pushing and deleting the identical image very
heavily in the e2e tests.

This experimental change enables setting the current time as the build
creation date on the image, which will lead to slightly different images
each build. At the moment it seems to be stopping the flake.

Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>