Use unique sources and builds in e2e tests
Google Artifact Registry has problems when pushing and deleting the same
blobs concurrently. This causes e2e flakes.

To work around this, we introduce unique.* files in the various source
directories and replace the occurences of the string `UNIQUENESS_TAG` with a GUID
when zipping the sources.

Now we zip the sources before each push rather than at the start of the
suite.

We have reverted the change which puts the current timestamp on the OCI
image. It is probably good to have the same SHA when pushing the same
sources, and this prevented that.

Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>