Remove verbose JSON from handlers app test
We now test the ForXXX methods in the presenter package. This checks the
entire JSON output.

In the handlers package, we just need to test enough to ensure the
correct ForXXX method has been invoked and the values have been passed
through.

This introduces a new MatchJSONPath gomega matcher which simplifies
asserting on fragments of JSON.

Issue: https://github.com/cloudfoundry/korifi/issues/2255
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>