Reorder buildpacks to make the cloud espm app deploy
The issue was that the gateway app has both a package.json and a
pom.xml. The old buildpack order used to detect it as a java app while
it is actually a node app. In traditional cf this is worked around by
specifying the builpack and skipping the detect phase.

This is a hack! Korifi should respect the specified buildpack
eventually.

Co-authored-by: Danail Branekov <danailster@gmail.com>