Avoid hanging org/space finalizers
When finalizing an org or space, the associated namespace is deleted,
and the finalizer is removed from the list. If the subsequent patch of
the metadata fails, we have a situation where the finalization will be
retried, but the namespace has already been deleted.

This commit ignores the now acceptable NotFound error so that
finalization can complete successfully.

(and fix an import warning in the space test)

Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>