Apps restart without downtime
We leverage the replicaset rollout capabilities to restart statefulset
pods. In order to achieve this, we implement `POST
/v3/apps/app-guid/actions/restart` to just put a
`korifi.cloudfoundry.org/restartedAt` annotaion which is propagated down
to the workload statefulset's `Spec.Template.ObjectMeta.Annotations`
Thise causes the replicaset controller to recrate the pods as their
template changed. FWIW, this is what `kubectl rollout restart
statefulset` does as well.

Things that can be improved: implement restat status condition (or
similar) propagation from the statefulset up to the cfapp so that the
restart operation can wait on that status condition.

Co-authored-by: Danail Branekov <danailster@gmail.com>